review case form functionality implemented
