+++ src/components/pages/Upload/UploadCase.js
@@ -1,43 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-// import UploadCaseForm from './UploadCaseForm';
 import { notification, Upload, Modal, Button, Spin } from 'antd';
 import {
   CheckCircleOutlined,
   CloseCircleOutlined,
   LoadingOutlined,
 } from '@ant-design/icons';
-import './CaseForm.css';
 import './_UploadCase.less';
 
 // Icons for modal
 import Icon from '@ant-design/icons';
 import UploadCaseBox from '../../../styles/icons/upload-box.svg';
 import OrangeLine from '../../../styles/orange-line.svg';
-
-// const initialFormValues = {
-//   date: '',
-//   judge: '',
-//   case_outcome: '',
-//   country_of_origin: '',
-//   protected_grounds: '',
-//   application_type: '',
-//   case_origin_city: '',
-//   case_origin_state: '',
-//   gender: '',
-//   applicant_language: '',
-//   indigenous_group: '',
-//   type_of_violence: '',
-//   initial_or_appellate: false,
-//   filed_in_one_year: false,
-//   credible: false,
-// };
-
 const UploadCase = ({ authState, getPendingCases }) => {
-  // const [formValues, setFormValues] = useState(initialFormValues);
-  const [formValueQueue, setFormValueQueue] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
-  const [isEditing, setIsEditing] = useState(false);
   const { Dragger } = Upload;
   const [postQueue, setPostQueue] = useState([]);
   const [nextPost, setNextPost] = useState(null);
@@ -98,12 +74,6 @@ const UploadCase = ({ authState, getPendingCases }) => {
     },
   };
 
-  // Might need this later depending on design changes
-  // const onInputChange = e => {
-  //   const { name, value } = e.target;
-  //   setFormValues({ ...formValues, [name]: value });
-  // };
-
   useEffect(() => {
     if (!nextPost && postQueue.length !== 0) {
       const copy = postQueue;
@@ -121,9 +91,7 @@ const UploadCase = ({ authState, getPendingCases }) => {
           },
         })
         .then(res => {
-          setFormValueQueue([...formValueQueue, res.data]);
           setIsLoading(false);
-          setIsEditing(true);
           successNotification();
           setNextPost(null);
           if (postQueue) {
@@ -205,14 +173,6 @@ const UploadCase = ({ authState, getPendingCases }) => {
               </form>
             </div>
           </div>
-          {
-            // Might need this later depending on design changes
-            /* <UploadCaseForm
-        formValues={formValues}
-        onInputChange={onInputChange}
-        formValueQueue={formValueQueue}
-      /> */
-          }
         </Modal>
       </div>
     </div>
