+++ src/components/pages/Upload/_UploadCase.less
@@ -77,3 +77,13 @@
 .divider {
   margin-bottom: 1.5rem;
 }
+
+.button-text {
+  color: white;
+}
+
+.spinner_container {
+  display: flex;
+  justify-content: center;
+  margin: 24px;
+}
